+++ test/test_models.py
@@ -59,6 +59,11 @@ class TestCollection:
 
         assert collection.registry_url is None
 
+    @pytest.mark.vcr
+    def test_get_archive_size(self):
+        collection = Collection.fetch('bigearthnet_v1_labels')
+        assert collection.archive_size == 173029030
+
 
 class TestDataset:
 
@@ -105,3 +110,9 @@ class TestDataset:
     def test_collections_list(self):
         dataset = Dataset.fetch('bigearthnet_v1')
         assert dataset.collections.__repr__() == '[<Collection id=bigearthnet_v1_source>, <Collection id=bigearthnet_v1_labels>]'
+
+    @pytest.mark.vcr
+    @pytest.mark.skip(reason="vcrpy does not handle multithreaded requests.")
+    def test_total_archive_size(self):
+        dataset = Dataset.fetch('bigearthnet_v1')
+        assert dataset.total_archive_size == 71311240007